requirements.txt
blinker==1.4
Flask==0.10.1
Flask-Login==0.3.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
SQLAlchemy==1.0.8
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
BeautifulSoup==3.2.1
Flask==0.10.1
Jinja2==2.8
Mako==0.9.1
MarkupSafe==0.18
PAM==0.4.2
Pillow==2.3.0
SPARQLWrapper==1.6.4
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
Werkzeug==0.10.4
apt-xapian-index==0.45
apturl==0.4.1ubuntu4
argparse==1.2.1
biopython==1.65
ccsm==0.9.11.3
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
compizconfig-python==0.9.11.3
configglue==1.1.2
configobj==4.7.2
debtagshw==0.1
defer==1.0.6
dirspec==13.10
dnspython==1.11.1
duplicity==0.6.23
feedparser==5.1.3
gyp==0.1
html5lib==1.0b8
httplib2==0.8
isodate==0.5.4
itsdangerous==0.24
lockfile==0.8
lxml==3.3.3
numpy==1.9.2
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
paramiko==1.10.1
pexpect==3.1
piston-mini-client==0.7.5
protobuf==2.5.0
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyinotify==0.9.4
pyparsing==2.0.3
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu1
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
radiotray==0.7.3
rdflib==4.2.2-dev
reportlab==3.0
requests==2.7.0
sessioninstaller==0.0.0
six==1.5.2
system-service==0.1.6
uTidylib==0.2
urlgrabber==3.9.1
urllib3==1.7.1
wsgiref==0.1.2
zope.interface==4.0.5